Don't bind the test to the current default, but to the desired behavior
we added a load statement to a before(:each) block in spot_rate_spec because
we need to reload the class to test default behavior that the gem
comes with out of the box, but we also want to test how the gem behaves
by default without being bound to our default currency converter (just
merely that whatever the default is, we'll use that).

Example: testing the .[] method we just want to ensure that whatever
the default is set to, that is used if no :use option key is specified,
but we don't particularily care what that default currency converter
is.

Conversely, we do want to ensure that the default currency
converter is set so anyone that downloads and uses the gem and
play-and-go without having to go through configuration steps or
authoring of their own converter.

Once we change the default converter, it stays changed until we reload
the class.